backend/requirements_minimal.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
setuptools>=65.0.0
chardet>=5.0.0
openpyxl>=3.1.0
pyarrow>=10.0.0
lxml>=4.9.0

# Essential document processing (reliable installs)
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
Pillow>=10.0.0

# Text processing
textblob>=0.17.1
requests>=2.31.0